+++ src/Octoshift/GithubApi.cs
@@ -481,32 +481,7 @@ namespace OctoshiftCLI
         {
             var url = $"{_apiUrl}/graphql";
 
-            var query = "query($id: ID!, $first: Int, $after: String)";
-            var gql = @"
-                node(id: $id) {
-                    ... on Organization {
-                        mannequins(first: $first, after: $after) {
-                            pageInfo {
-                                endCursor
-                                hasNextPage
-                            }
-                            nodes {
-                                login
-                                id
-                                claimant {
-                                    login
-                                    id
-                                }
-                            }
-                        }
-                    }
-                }";
-
-            var payload = new
-            {
-                query = $"{query} {{ {gql} }}",
-                variables = new { id = orgId }
-            };
+            var payload = GetMannequinsPayload(orgId);
 
             var mannequin = await _client.PostGraphQLWithPaginationAsync(
                     url,
@@ -515,20 +490,24 @@ namespace OctoshiftCLI
                     data => (JObject)data["data"]["node"]["mannequins"]["pageInfo"])
                 .FirstOrDefaultAsync(jToken => username.Equals((string)jToken["login"], StringComparison.OrdinalIgnoreCase));
 
-            return mannequin is null
-                ? new Mannequin()
-                : new Mannequin
-                {
-                    Id = (string)mannequin["id"],
-                    Login = (string)mannequin["login"],
-                    MappedUser = mannequin["claimant"].Any()
-                    ? new Claimant
-                    {
-                        Id = (string)mannequin["claimant"]["id"],
-                        Login = (string)mannequin["claimant"]["login"]
-                    }
-                    : null
-                };
+            return mannequin is null ? new Mannequin() : MannequinBuilder(mannequin);
+        }
+
+        public virtual async Task<IEnumerable<Mannequin>> GetMannequins(string orgId)
+        {
+            var url = $"{_apiUrl}/graphql";
+
+            var payload = GetMannequinsPayload(orgId);
+
+            return await _client.PostGraphQLWithPaginationAsync(
+                    url,
+                    payload,
+                    data => (JArray)data["data"]["node"]["mannequins"]["nodes"],
+                    data => (JObject)data["data"]["node"]["mannequins"]["pageInfo"])
+                .Select(mannequin => MannequinBuilder(mannequin)
+
+                )
+                .ToListAsync();
         }
 
         public virtual async Task<string> GetUserId(string login)
@@ -580,6 +559,52 @@ namespace OctoshiftCLI
             var data = JObject.Parse(response);
 
             return data.ToObject<MannequinReclaimResult>();
+        }
+
+        private static object GetMannequinsPayload(string orgId)
+        {
+            var query = "query($id: ID!, $first: Int, $after: String)";
+            var gql = @"
+                node(id: $id) {
+                    ... on Organization {
+                        mannequins(first: $first, after: $after) {
+                            pageInfo {
+                                endCursor
+                                hasNextPage
+                            }
+                            nodes {
+                                login
+                                id
+                                claimant {
+                                    login
+                                    id
+                                }
+                            }
+                        }
+                    }
+                }";
+
+            return new
+            {
+                query = $"{query} {{ {gql} }}",
+                variables = new { id = orgId }
+            };
+        }
+
+        private static Mannequin MannequinBuilder(JToken mannequin)
+        {
+            return new Mannequin
+            {
+                Id = (string)mannequin["id"],
+                Login = (string)mannequin["login"],
+                MappedUser = mannequin["claimant"].Any()
+                                    ? new Claimant
+                                    {
+                                        Id = (string)mannequin["claimant"]["id"],
+                                        Login = (string)mannequin["claimant"]["login"]
+                                    }
+                                    : null
+            };
         }
     }
 }